Writes 0xffffff00 to APPROTECT instead of flashing a bin file
OpenOCD reset required for the nRF52 to read in UICR registers for any changes to take effect.  Or you can power cycle the nRF52.
Shutting down and powering off your Raspberry Pi assumes you are powering the nRF52 from the Raspberry Pi and there is not other supply to the nRF52.